Add Clash.Normalize.Transformations.Reduce
All transformations concerning reduction now live in a separate
module. This is part of work to make the normalization modules
quicker to compile.
